Don't perform now-invalid cast to `usize`
In response to briansmith/ring@8a655e45fc0e7926db3932677a82e875bc559ce6.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.

+++ ring/ring.rs
@@ -115,13 +115,13 @@ mod pbkdf2 {
 
     pbkdf2_bench!(hmac_sha256, crypto_bench::SHA256_OUTPUT_LEN, out,
                   pbkdf2::derive(&pbkdf2::HMAC_SHA256,
-                                 crypto_bench::pbkdf2::ITERATIONS as usize,
+                                 crypto_bench::pbkdf2::ITERATIONS,
                                  &crypto_bench::pbkdf2::SALT,
                                  crypto_bench::pbkdf2::PASSWORD, &mut out));
 
     pbkdf2_bench!(hmac_sha512, crypto_bench::SHA512_OUTPUT_LEN, out,
                   pbkdf2::derive(&pbkdf2::HMAC_SHA512,
-                                 crypto_bench::pbkdf2::ITERATIONS as usize,
+                                 crypto_bench::pbkdf2::ITERATIONS,
                                  crypto_bench::pbkdf2::SALT,
                                  crypto_bench::pbkdf2::PASSWORD, &mut out));
 }
